somewhat decent progress for multi loader environment. missing configs
