+++ build.gradle
@@ -1,101 +1,92 @@
 plugins {
-	id 'fabric-loom' version '0.12-SNAPSHOT'
-	id 'maven-publish'
+    id "architectury-plugin" version "3.4-SNAPSHOT"
+    id "dev.architectury.loom" version "0.12.0-SNAPSHOT" apply false
 }
 
-sourceCompatibility = JavaVersion.VERSION_17
-targetCompatibility = JavaVersion.VERSION_17
-
-archivesBaseName = project.archives_base_name
-version = "${project.mod_version}+${project.minecraft_version}"
-group = project.maven_group
-
-repositories {
-	mavenLocal()
-
-	// Mod Menu.
-	maven { url 'https://maven.terraformersmc.com' }
-
-	// MidnightControls.
-	maven {
-		name 'Gegy'
-		url 'https://maven.gegy.dev'
-	}
-	maven { url = "https://aperlambda.github.io/maven" }
-	maven { url = "https://api.modrinth.com/maven" }
-
-	// Cloth Config.
-	maven { url "https://maven.shedaniel.me/" }
-	maven { url "https://maven.terraformersmc.com/releases/" }
-
-	// Jitpack for Mixin Extras.
-	maven { url 'https://jitpack.io' }
+architectury {
+    minecraft = rootProject.minecraft_version
 }
 
-dependencies {
-	minecraft "com.mojang:minecraft:${project.minecraft_version}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
-	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
-	modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
+subprojects {
+    apply plugin: "dev.architectury.loom"
 
-	// Mixin Extras for extra compatibility.
-	include(implementation("com.github.LlamaLad7:MixinExtras:${project.mixin_extras_version}"))
-	annotationProcessor("com.github.LlamaLad7:MixinExtras:${project.mixin_extras_version}")
+    loom {
+        silentMojangMappingsLicense()
+    }
 
-	// Mod Menu integration. Currently unused.
-	modImplementation "com.terraformersmc:modmenu:${project.modmenu_version}"
+    dependencies {
+        minecraft "com.mojang:minecraft:${rootProject.minecraft_version}"
+        // The following line declares the mojmap mappings, you may use other mappings as well
+        mappings loom.layered() {
+            //it.addLayer(quiltMappings.mappings("org.quiltmc:quilt-mappings:${rootProject.quilt_version}"))
+            it.parchment("org.parchmentmc.data:parchment-${rootProject.parchment_version}")
+            it.officialMojangMappings {
+                setNameSyntheticMembers(false)
+            }
+        }
+        // The following line declares the yarn mappings you may select this one as well.
+        // mappings "net.fabricmc:yarn:1.19+build.4:v2"
 
-	// Cloth Config for easy config menu.
-	modApi("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version}") {
-		exclude group: "net.fabricmc.fabric-api"
-	}
-
-	// Controller Support
-	modImplementation "dev.lambdaurora:spruceui:${project.spruceui_version}"
-	modImplementation "maven.modrinth:midnightlib:${project.midnightlib_version}"
-	modImplementation "maven.modrinth:midnightcontrols:${project.midnightcontrols_version}"
-	api('org.aperlambda:lambdajcommon:1.8.1') {
-		exclude group: 'com.google.code.gson'
-		exclude group: 'com.google.guava'
-	}
+        // Mixin Extras for extra compatibility.
+        include(implementation("com.github.LlamaLad7:MixinExtras:${project.mixin_extras_version}"))
+        annotationProcessor("com.github.LlamaLad7:MixinExtras:${project.mixin_extras_version}")
+    }
 }
 
-processResources {
-	inputs.property "version", project.version
+allprojects {
+    apply plugin: "java"
+    apply plugin: "architectury-plugin"
+    apply plugin: "maven-publish"
 
-	filesMatching("fabric.mod.json") {
-		expand "version": project.version
-	}
-}
+    archivesBaseName = rootProject.mod_id
+    version = rootProject.mod_version
+    group = rootProject.maven_group
 
-tasks.withType(JavaCompile).configureEach {
-	it.options.encoding = "UTF-8"
-	it.options.release = 17
-}
+    repositories {
 
-java {
-	withSourcesJar()
-}
+        flatDir {
+            dirs 'mods'
+        }
 
-jar {
-	from("LICENSE") {
-		rename { "${it}_${project.archivesBaseName}"}
-	}
-}
+        mavenLocal()
 
-publishing {
-	publications {
-		mavenJava(MavenPublication) {
-			artifact(remapJar) {
-				builtBy remapJar
-			}
-			artifact(sourcesJar) {
-				builtBy remapSourcesJar
-			}
-		}
-	}
+        // Add repositories to retrieve artifacts from in here.
+        // You should only use this when depending on other mods because
+        // Loom adds the essential maven repositories to download Minecraft and libraries from automatically.
+        // See https://docs.gradle.org/current/userguide/declaring_repositories.html
+        // for more information about repositories.
+        maven { url "https://maven.parchmentmc.org" }
 
-	repositories {
-		
-	}
+        maven {
+            url = "https://www.cursemaven.com"
+            allowInsecureProtocol = true
+        }
+
+        // Mod Menu.
+        maven { url 'https://maven.terraformersmc.com' }
+
+        // MidnightControls.
+        maven {
+            name 'Gegy'
+            url 'https://maven.gegy.dev'
+        }
+        maven { url = "https://aperlambda.github.io/maven" }
+        maven { url = "https://api.modrinth.com/maven" }
+
+        // Cloth Config.
+        maven { url "https://maven.shedaniel.me/" }
+        maven { url "https://maven.terraformersmc.com/releases/" }
+
+        // Jitpack for Mixin Extras.
+        maven { url 'https://jitpack.io' }
+    }
+
+    tasks.withType(JavaCompile) {
+        options.encoding = "UTF-8"
+        options.release = 17
+    }
+
+    java {
+        withSourcesJar()
+    }
 }
