+++ common/src/main/java/nl/enjarai/doabarrelroll/config/ConfiguresRotation.java
