+++ common/src/main/java/nl/enjarai/doabarrelroll/mixin/ClientPlayerEntityMixin.java
@@ -1,0 +1,34 @@
+package nl.enjarai.doabarrelroll.mixin;
+
+import com.mojang.authlib.GameProfile;
+import net.minecraft.client.ClientRecipeBook;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.multiplayer.ClientPacketListener;
+import net.minecraft.client.player.AbstractClientPlayer;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.stats.StatsCounter;
+import net.minecraft.world.entity.player.ProfilePublicKey;
+import nl.enjarai.doabarrelroll.DoABarrelRollClient;
+import nl.enjarai.doabarrelroll.ElytraMath;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(LocalPlayer.class)
+public abstract class ClientPlayerEntityMixin extends AbstractClientPlayer {
+
+	@Shadow public abstract float getViewYRot(float tickDelta);
+
+	public ClientPlayerEntityMixin(ClientLevel world, GameProfile profile, ProfilePublicKey publicKey) { super(world, profile, publicKey); }
+
+	@Inject(
+			method = "<init>",
+			at = @At("RETURN")
+	)
+	public void doABarrelRoll$init(Minecraft client, ClientLevel world, ClientPacketListener networkHandler, StatsCounter stats, ClientRecipeBook recipeBook, boolean lastSneaking, boolean lastSprinting, CallbackInfo ci) {
+		DoABarrelRollClient.left = ElytraMath.getAssumedLeft(getYRot());
+	}
+}
