+++ common/src/main/resources/architectury.common.json
@@ -1,0 +1,3 @@
+{
+  "accessWidener": "do-a-barrel-roll.accesswidener"
+}
