+++ common/src/main/resources/assets/do-a-barrel-roll/icon.png
