+++ common/src/main/resources/do-a-barrel-roll.accesswidener
@@ -1,0 +1,1 @@
+accessWidener   v2  named
