+++ common/src/main/resources/do-a-barrel-roll.compat.midnightcontrols.mixins.json
