+++ fabric/src/main/java/nl/enjarai/doabarrelroll/config/fabric/ModConfigImpl.java
@@ -1,0 +1,7 @@
+package nl.enjarai.doabarrelroll.config.fabric;
+
+public class ModConfigImpl {
+    public static boolean getModEnabled() {
+        //cloth config stuff here. for yacl i only ever used the menu part so idk
+    }
+}
