+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,50 @@
+{
+	"schemaVersion": 1,
+	"id": "do-a-barrel-roll",
+	"version": "${version}",
+	
+	"name": "Do a Barrel Roll",
+	"description": "Microsoft flight simulator for Minecraft elytras.",
+	"authors": [
+		"enjarai",
+		"Jorbonism",
+		"Motschen"
+	],
+	"contact": {
+		"homepage": "https://enjarai.nl",
+		"sources": "https://github.com/enjarai/do-a-barrel-roll"
+	},
+	
+	"license": "GPL-3.0",
+	"icon": "assets/do-a-barrel-roll/icon.png",
+
+	"entrypoints": {
+		"client": [
+			"nl.enjarai.doabarrelroll.DoABarrelRollClient"
+		],
+		"modmenu": [
+          "nl.enjarai.doabarrelroll.compat.modmenu.ModMenuIntegration"
+		],
+		"preLaunch": [
+			"nl.enjarai.doabarrelroll.DoABarrelRollPreLaunch"
+		]
+	},
+	
+	"environment": "client",
+	"mixins": [
+		"do-a-barrel-roll.mixins.json",
+		"do-a-barrel-roll.compat.midnightcontrols.mixins.json"
+	],
+
+	"depends": {
+		"cloth-config": "*",
+		"minecraft": "1.19.*"
+	},
+
+	"custom": {
+		"projects": {
+			"modrinth": "6FtRfnLg",
+			"curseforge": 663658
+		}
+	}
+}
