+++ gradle.properties
@@ -1,17 +1,34 @@
-org.gradle.jvmargs=-Xmx1G
+org.gradle.jvmargs=-Xmx4096M
 
-minecraft_version=1.19
-yarn_mappings=1.19+build.4
-loader_version=0.14.8
+minecraft_version = 1.19.2
+enabled_platforms = fabric,forge
 
-mod_version = 1.3.3
-maven_group = nl.enjarai
-archives_base_name = do-a-barrel-roll
+mod_id = do-a-barrel-roll
+mod_version = 1.19-1.3.3
+maven_group =  nl.enjarai
 
-fabric_version=0.58.0+1.19
-modmenu_version=4.0.4
+parchment_version = 1.19.2:2022.08.10
+quilt_version = 1.18.2+build.24:v2
+
+fabric_loader_version=0.14.9
+fabric_api_version=0.59.0+1.19.2
+
+forge_version=1.19.2-43.1.30
+
+modmenu_version = 4.0.4
+cloth_version = 7.0.72
+
+
 cloth_config_version=8.0.75
 spruceui_version=4.0.0+1.19
 midnightlib_version=0.5.2
 midnightcontrols_version=1.5.0+1.19
 mixin_extras_version=0.1.0-rc5
+
+
+
+
+
+
+
+
