--- src/main/java/nl/enjarai/doabarrelroll/DoABarrelRollPreLaunch.java
@@ -1,11 +1,0 @@
-package nl.enjarai.doabarrelroll;
-
-import com.llamalad7.mixinextras.MixinExtrasBootstrap;
-import net.fabricmc.loader.api.entrypoint.PreLaunchEntrypoint;
-
-public class DoABarrelRollPreLaunch implements PreLaunchEntrypoint {
-    @Override
-    public void onPreLaunch() {
-        MixinExtrasBootstrap.init();
-    }
-}
